pyproject.toml
[tool.black]
line-length = 120
target_version = ['py39']
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | \.vscode
)
'''

[tool.isort]
profile = "black"

[tool.poetry]
name = "python-telegram-bot-django-persistence"
version = "0.1.6"
description = "Package to use Django ORM as persistence engine in Python Telegram Bot"
authors = ["Alexander Shishenko <alex@shishenko.com>"]
license = "MIT"
classifiers = [
    'Intended Audience :: Developers',
    'Topic :: Internet',
]
readme = "README.md"
homepage = "https://github.com/GamePad64/python-telegram-bot-django-persistence"

[tool.poetry.dependencies]
python = "^3.9"
Django = ">=3.1"
python-telegram-bot = ">=13.7"

[tool.poetry.dev-dependencies]
black = ">=21.7b0"
isort = "^5.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
